b532: prevent_delete_executions

+++ implementation/execution/user.go
@@ -7,6 +7,7 @@ import (
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	"os/exec"
+	"runtime"
 	"strings"
 )
 
@@ -34,6 +35,7 @@ func Sanitize(username string) string {
 
 func AddNewUser(u *User) (string, error) {
 
+	log.Infof("adding new user named %s", u.Name)
 	passwordBase, err := CreateRandom(randomPasswordLength)
 	if err != nil {
 		return "", err
@@ -63,6 +65,10 @@ func AddNewUser(u *User) (string, error) {
 }
 
 func RemoveUser(username string) error {
+	if runtime.GOOS != "linux" {
+		return nil
+	}
+
 	argUser := []string{"-r", username}
 	userCmd := exec.Command("userdel", argUser...)
